Clean dirStat when uploading directories
When uploading dirs to a different box and/or OS, the dirstat can
contain an owner/group ids that aren't available at the destination.

This can result in PERMISSION DENIED errors when the target user tries
to create files with an incorrect u(g)id.

This change only propagates file mode to the destination. The files
will be owned by the remote user.